requirements.txt
pre_commit==4.5.0
python-dotenv==1.2.1
